pyproject.toml
[project]
name = "luminaai"
version = "0.1.0"
description = ""
authors = [
    {name = "Gaurav Chaudhari",email = "gauravchaudhari37@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit (>=1.42.2,<2.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "numpy (>=2.2.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
